Move cohort engine CLI to separate project, implement measure CLI
