+++ cohort-cli/config/fhir-server.cer
