+++ cohort-cli/src/test/resources/cql/basic/test.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<library xmlns="urn:hl7-org:elm:r1" xmlns:t="urn:hl7-org:elm-types:r1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:fhir="http://hl7.org/fhir" xmlns:qdm43="urn:healthit-gov:qdm:v4_3" xmlns:qdm53="urn:healthit-gov:qdm:v5_3" xmlns:a="urn:hl7-org:cql-annotations:r1">
+   <annotation translatorOptions="" xsi:type="a:CqlToElmInfo"/>
+   <identifier id="Test" version="1.0.0"/>
+   <schemaIdentifier id="urn:hl7-org:elm" version="r1"/>
+   <usings>
+      <def localIdentifier="System" uri="urn:hl7-org:elm-types:r1"/>
+      <def localIdentifier="FHIR" uri="http://hl7.org/fhir" version="4.0.0"/>
+   </usings>
+   <contexts>
+      <def name="Patient"/>
+   </contexts>
+   <statements>
+      <def name="Patient" context="Patient">
+         <expression xsi:type="SingletonFrom">
+            <operand dataType="fhir:Patient" xsi:type="Retrieve"/>
+         </expression>
+      </def>
+      <def name="Female" context="Patient" accessLevel="Public">
+         <expression xsi:type="Equal">
+            <operand path="value" xsi:type="Property">
+               <source path="gender" xsi:type="Property">
+                  <source name="Patient" xsi:type="ExpressionRef"/>
+               </source>
+            </operand>
+            <operand valueType="t:String" value="female" xsi:type="Literal"/>
+         </expression>
+      </def>
+      <def name="Male" context="Patient" accessLevel="Public">
+         <expression xsi:type="Equal">
+            <operand path="value" xsi:type="Property">
+               <source path="gender" xsi:type="Property">
+                  <source name="Patient" xsi:type="ExpressionRef"/>
+               </source>
+            </operand>
+            <operand valueType="t:String" value="male" xsi:type="Literal"/>
+         </expression>
+      </def>
+      <def name="Over the hill" context="Patient" accessLevel="Public">
+         <expression xsi:type="GreaterOrEqual">
+            <operand precision="Year" xsi:type="CalculateAge">
+               <operand path="birthDate.value" xsi:type="Property">
+                  <source name="Patient" xsi:type="ExpressionRef"/>
+               </operand>
+            </operand>
+            <operand valueType="t:Integer" value="40" xsi:type="Literal"/>
+         </expression>
+      </def>
+   </statements>
+</library>
+
+
