+++ cohort-cli/src/test/resources/cql/multi-folder/FHIRHelpers.cql
