+++ cohort-cli/src/test/resources/cql/multi-folder/cohort1/Cohort1-1.0.0.cql
