+++ cohort-cli/src/test/resources/cql/multi-folder/cohort2/Cohort2-1.0.0.cql
