+++ cohort-cli/src/test/resources/cql/zip-with-folders/cohorts.zip
