+++ cohort-engine/src/main/java/com/ibm/cohort/engine/measure/MeasureEvaluator.java
@@ -13,6 +13,7 @@ import org.hl7.fhir.r4.model.Measure;
 import org.hl7.fhir.r4.model.MeasureReport;
 import org.opencds.cqf.common.evaluation.EvaluationProviderFactory;
 import org.opencds.cqf.common.providers.LibraryResolutionProvider;
+import org.opencds.cqf.cql.engine.debug.DebugMap;
 import org.opencds.cqf.cql.engine.execution.LibraryLoader;
 import org.opencds.cqf.r4.evaluation.MeasureEvaluationSeed;
 
@@ -71,6 +72,11 @@ public class MeasureEvaluator {
 		// configuration of our FHIR server.
 		seed.setup(measure, periodStart, periodEnd, /* productLine= */"ProductLine", /* source= */"", /* user= */"",
 				/* pass= */"");
+		
+		// This is enabled by default in cqf-ruler and we don't want it.
+		DebugMap debugMap = new DebugMap();
+		debugMap.setIsLoggingEnabled(false);
+		seed.getContext().setDebugMap(debugMap);
 
 		// TODO - The OSS logic takes converts the period start and end into an
 		// Interval and creates a parameter named "Measurement Period" that is populated
