+++ cohort-parent/pom.xml
@@ -12,6 +12,7 @@
 
 	<modules>
 		<module>../cohort-engine</module>
+		<module>../cohort-cli</module>
 		<module>../cdm-modelinfo-tooling</module>
 		<module>../tests</module>
 	</modules>
