+++ docs/pom.xml
@@ -1,3 +1,6 @@
+<!-- 
+    This FILE is ONLY so the files may be IMPORTED into Eclipse. 
+ -->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
